Don't kill an ancestor every time the player takes the stairs (Shummie)
After the recent refactoring, the simplified ally dismissal code was
killing *all* monsters on the floor the player was leaving that were marked
as having been created by the player. And for some reason, Hep ancestors
considered the player to be their summoned.

I don't *think* there will be any issue with them just not having a
summoner marked at all, and this fixes the present issue. (If some reason
is discovered why this marking is important, I can fix it in some other
way.)