Don't crash on monsters dying to dead monsters (+fix monster frenzy orb)
The guard which prevented a crash on kills by dead monsters (such as
exploders) was removed by 400f974 at least in part because the comment
above it was misleading enough that I thought it wasn't necessary.

This also fixes a second bug among the same code, where Orbs of Mayhem
wielded by monsters could only trigger if the player currently worshipped
TSO and the monster which died was evil. ...because for some reason, the
mayhem check was in the TSO follower blessing codeblock, and appears to
have always been.