Fix Makhleb god description not rending on webtiles
fc8bfd6970fa08 caused the description + cost of Makhleb's capstone ability
to be exactly 80 characters in length, which meant there was no space
between the description and the cost (longer strings get spaces inserted,
but *exatly* 80 characters does not). This caused the regex the webtiles
UI uses to split the string to not work.

This commit changes the regex to no longer require a leading space before
the parantheses, since those are only used for costs here anyway.

This fixes #4068