Unify handle of channelled spells
Searing Ray, Flame Wave, and Maxwell's Capacitive Coupling all had a bunch
of common behavior (ie: ending on movement or becoming silenced during the
casting of the spell) that each spell implemented individually - and
sometimes in slightly different ways from each other.

This is an attempt to provide common functions to handle these shared
behaviors in a unified way (and make it easier to add a 4th such spell
without more code duplication).

This doesn't clean up all props used by these spells the way the old code
did, but that shouldn't matter - the props will always be set to relevant
values when the spell is cast, and ignored when the spell is not active.

This commit also contains save compat code to handle anyone who happened to
have saved in the middle of channelling one of these spells.