New creeping inferno tile (pianoman523)
As with the apis change in b8f47f2 and retiling in b29aa79, this
submission was delayed until the mechanics for the monster were reworked,
to hopefully help tiles players associate a new appearance with a
drastically different behaviour. This combines and edits multiple
submissions by the same artist; it still looks a bit silly, but vastly
less janky than the previous creeping inferno's lurking horror recolour
tile.