Name removed spells when failing to learn or cast them
It can be quite annoying to type in wizmode e.g. just "Conjure" for Ball
Lightning and not be explicitly told it first found Conjure Flame, or
that "Summon S" fills in the rest as Swarm and not Small Mammals, so
we might as well list what exactly it tried to find instead.