Tweak confusion's interaction with tension
Confusion is easily cured with potions of curing later in the game, so
staying in confusion being a high multiplier of tension is sort of weird
in so far as one would only do such if they weren't any concerned with
the encounter. (However, it's still plenty lethal very early on, when
one doesn't know about potions of curing yet.) Contrariwise, confusion
is substantially worse than many other statuses if one can't actually
cure it any time soon due to not being able to drink potions (whether
in Cocytus, while possessing -Potion status, or being a mummy / lich).

As such, this takes the current x2 tension multiplier for confusion,
sets it slightly higher by default, and drops that slowly down versus
one's current XL, unless one can't drink to cure it which case it
becomes a flat x2.5 multiplier. (This is a little silly for Elyvilon, but
there's a piety cost involved for the extremely little non-Xom non-wrath
uses of tension that could even come up versus -Potion.) Hopefully this
should help out Mummy Chaos Knights a little bit more and get tension the
littlest bit closer to being a more usable function.