Spirit Yak: new monster
The venerable psychopomp of the yaktaur afterlife, it herds the souls of
the deceased into the next life.

As an expansion of the yaktaur monster set, it has a strong support aura
that specifically benefits ranged monsters. This bestows eligible allied
monsters with "spirit arrows", giving ranged weapons an additional free
shot at the same target, only if the first hit is successful hit. Being
non corporeal, this shot is able to bypass half the target's AC.

Because the spirit arrow only processes on a connecting hit, players
with good EV or SH may fare rather better than those relying more
heavily on AC, and while it bypasses some AC it can still be blocked or
evaded like a regular missile.

It has a 50% chance to appear in Yaktaur Cleric bands to fit their
necromantic flavour, and can additionally be found wandering in the wild
with its *own* band set of recently-deceased archers (not necessarily
yaktaurs) on their way to their next life.

The tile is a hastily spectralised version of the Plumootoo's "pretty
yak" and could probably be improved by an artist.

Still to do:
- Strip brands from the spirit arrow