Fixed 0_stat effects for revived characters.
There are duration counters to track how long Collapse should hold. If strength > 0 the counter is decreased, and Collapse is removed when the counter is 0. On reviving the counter was mistakenly set to zero. Same for Brainless and Clumsy.

Now revive does not set those counters to 0, to retain the 0_stat
effects across being revived.