New spell: Construct Spike Launcher (Level 2 Forgecraft)
Another Forgewright starter spell, this temporarily transforms a random
adjacent rock wall into a 'trap'/turret that repeatedly attacks a random
adjacent enemy each turn. If the caster moves more than 2 tiles away from
this launcher, it will disappear prematurely.

The idea here was to explore creating an 'ally' that couldn't be used as
an earlygame corridor swap or obstruction for enemies (which Summoner is
famously good at). In practice, it plays a little like making a single-tile
and single-target Frozen Ramparts, which allows a bit more movement while
active (and might, in fact, require the caster to move immediately in order
to position an enemy in the launcher's range). But that is a popular and
successful spell, and I feel this is more than sufficiently distinct from
it.

(I would have been happy to have this spell by targeted as Forgewright's
level 1 spell, but its terrain-dependence means that some entry vaults
simply will not allow the player to use it at all in their initial battles
and that feels unacceptable.)