New spell: Kinetic Grapnel (Level 1 Forgecraft)
Given that Spike Launcher will be online for most Forgewrights by XL 2,
this is a spell with a short shelf-life, but they *do* need a generally
reliable XL 1 option and this is intended to be that.

Kinetic Grapnel is a short-range direct-damage projectile. If it
successfully does damage, it lodges a grapnel in that enemy which allows
the caster's next melee attack against them to be an automatic hit with +3
slaying.

Requiring a dodgeable, low-damage projectile to do damage should hopefully
limit its long-term usefulness as an accuracy boost (and the flat slaying
falls off quickly), but this forefronts some of the 'melee hybrid'
properties of Forgecraft from XL 1 while offering a straightforward tool
before more positional ones come online.

(The spell tiles at the moment don't look very grapnel-like, but this was
at least the 3rd level 1 spell I tried, so we'll just consider them a
placeholder for now.)