New spell: Diamond Sawblades (level 7 Forgecraft)
This spell creates up to 4 sawblades in the following fixed positions
around the player, which deal damage to adjacent enemies each turn they
are active:

X...X
.....
..@..
.....
X...X

(They cannot be placed in any space the player cannot see, or one with a
wall or other monster.)

The sawblades can be used both as AoE damage, or short-lived obstructions
(though they're not terribly durable) and the unique targeting will
hopefully interact with different spaces in interesting ways. (For
instance, they are useless in straight corridors, but very strong in
diagonal ones, and can be used both defensively and offensively in a single
cast, with the proper positioning.)

They are intended as a flexible and straightforward high-level
single-school option.