Changed when brand_damage is done during attacking
Moved brand damage calculation from phase_hit to phase_damaged.
The reason for this is to have `apply_damage_brand()` called right
after `announce_hit()`, before `mons_attack_effects()` is called. The
reason is that both methods are closely linked to the weapon, while
`mons_attack_effects()` is not.

This reordering has the benefit that it fixes #4041, where
`apply_damage_brand()` was called when the attacking monster
was shafted in the same turn it attacked.

The bug could also have been fixed by simply checking inside
`apply_damage_brand()`, whether the attacker is still alive. However,
that route seemed less sensible to me.

I also moved the call to `_apply_flux_contam()` because in its nature
it seemed quite similar to a weapon brand.