Fix Wucad Mu / Borg wizlab summons being broken (midn8)
When summon mgen_data was switched to using aut durations instead of abj
degree, vaults which manually created summoned monsters was overlooked.
Since it's a bit of a pain to do the per-monster duration randomization in
vaultspec, I've made "dur:X" there default to calling summ_dur directly.

This means it isn't possible to specify summon duration in aut in vaults
(or wizmode monster creation...), which I consider inideal, but this is a
reasonable quick fix given how rarely this functionality is used.

(This also makes all summoned monsters made by monspec 'actual' summons,
which isn't strictly true in theory, but is true for all current uses I
know of. Slightly unhappy about this.)