Fix directional bias in pool fixup
Due to the algorithm modifying tiles as it went, subsequent iterations
were based on the new state rather than the original - this resulted
in North- and West-facing pools getting affected potentially several
tiles deep, while the South and East edges general got just a one tile
layer of shallow water, which looked a bit odd and not symmetrical.

The adjusted algorithm builds a vector of coords to paint later: however
this looked rather too neat as then there is only ever a 1-tile border
of shallow water. So I ran a 2nd pass to let the shallow water encroach
to 2 tiles deep in places. The probabilities needing tweaking to
produce a satisfying organic look that is quite varied.

Sometimes the 2nd pass will not run, with a 25% chance, to produce
tidier pools and moats occasionally (but not in Lair).