Fix randomness in shield block to-hit calculation
To account for shield block in hit %, shield_bonus() was copied to
monster_info; however slightly unexpectedly there was a random roll
being applied in this method. This meant that for monsters with shields
(or an amulet of reflection) the to-hit % would randomly change each
time inspected.

Additionally the way integer division was used across these methods
meant there was some further inaccuracy in the simple average calculated
for some values of sh.

This change improves the formula a bit; it's still not perfect but
definitely better. It also now correctly indicates when a monster's
incapacitated status is affecting the shield block, if that monster had
zero dodge normally.