Add the Abjuration spell to clua monster info
Monsters with spells marked with the summon flag get Abjuration the
automatically, although only cast with a chance when they would cast a
summon spell. This special case has to be handled separately if we're
listing a monster's spells. Hence this commit adds Abjuration to a
monster's spell list in clua when appropriate.