New spell: Eringya's Surprising Crocodile (Level 4 Summonings)
Intended as a Summoner starting to spell to replace the two level 4s it's
about to lose to Forgewright.

This spell summons a crocodile as a short-term mount beneath the caster,
which performs an ambush attack and repositions the caster in a single
action.

This spell is cast upon a target in melee range. The summoned crocodile
will perform an immediate attack against that enemy (which never misses and
has bonus damage based on spellpower), and drag both it and the caster
backwards. Then the caster hops off the crocodile, landing a space further
back, while the crocodile remains as a normal summon. It also makes some
water along the drag path.

The spell cannot be recast while the crocodile is still active (so you
can't perform silly chain-grabs), but it's fairly short-lived and fragile.

This is intended to be a more dynamic use of summons than average, and
allow some useful repositioning - leaning into crocodile's natural drag
attack by making it even better at deliberately pulling enemies away from
their allies to surround them with your own creatures, while also putting
the crocodile between you and whatever enemy was just next to you.