Properly remove Animate Skeleton
Animate Skeleton was nominally removed in 613369f but the removal was
somewhat incomplete. The spell description was left in a number of
translations (just removed in en), the spell tile still remained,
the spell data was still present and not made into an AXED_SPELL, the
enum was not marked for TAG_MAJOR_VERSION deletion, and (consequently)
it was not added to removed_spells.

So to avoid potential confusion and errors I have done the above. I
also added a fixup in tags.cc so very old Necromancer saves should now
receive Soul Splinter instead (which it was ultimately replaced with).