Changed resting with inhibited regeneration mutation
So far, resting with the mutation inhibited regeneration while a
invisible monster is nearby that the player is not aware off resulted in
the normal "You start resting" message, without letting any time pass.

Now, a message is generated that indicates that there is a monster
nearby, which inhibits regeneration, and still we do not let time pass.

Additionally, changed the description of the mutation to clearly state
that monsters need not be visible to prevent regeneration.