Fix inhibited regen infoleaking invis mons
Previously, the inhibited regen mutation effect was being applied while
any monster was nearby, regardless of whether the player knew that it was
there. This was technically an infoleak, and also led to messaging that
players found confusing: E.g. with a nearby unseen horror, players couldn't
rest, and they weren't being informed why in the combat log.

This commit changes the behavior of inhibited regen to only take effect
with visible nearby monsters: there shouldn't be any issues with this from
a gameplay perspective, and this fixes both the infoleak and the messaging.

Fixes #4018.
Closes #4064.