New spell: Platinum Paragon (level 9 Forgecraft)
The capstone of the Forgecraft school, this spell is intended as an
ally-creation spell that holds up as being worthly of level 9, while also
being as different from Dragon's Call as is possible.

Instead of creating a large number of short-lived, disposable, regular
monsters, it creates a single unique ally that you are encouraged to keep
alive (at least for a little while) and even fight beside (as opposed to
maneuvering to use your dragon swarm to *prevent* enemies from being
beside you.)

As a single bespoke ally with modality, Paragon has a lot of things going
on, but I think many of these properties flow naturally together and play
in a fun and distinctive way:

-Your Paragon can wield any melee weapon the player has found. (You use the
 Imprint Weapon ability to select this weapon, similar to how Imbue
 Servitor is used to select servitor spell.) Whenever you cast Paragon
 after this, it will spawn wielding a perfect replica of that weapon (so
 there's no need to carry it around in your inventory or switch to it
 before casting the spell).
-Whenever the caster performs a normal melee action, your paragon will also
 perform an instant melee attack again a random enemy in its range.
-While within 2 tiles of your paragon, it has a 1-in-3 chance to block any
 attack targeting the player. (This is aimed to further encourage the
 caster, who might be themselves frail to have a level 9 spell, to melee
 alongside their paragon with more confidence.)
-The paragon builds up charge as it melees things, and when it has attained
 a sufficient amount, the spell can be recast to perform a powerful
 finisher - unleashing a weapon attack with a bonus damage multiplier again
 many nearby enemies, and destroying the paragon in the process.
-Casting the spell itself deploys the paragon to a selected nearby location
 and deals modest damage via a concussive shockwave to nearby enemies
 (aimed at making the cycle of 'deliberately destroy paragon, replace
 paragon' feel better since you also get an immediate benefit to making a
 new one.)

All together, this makes a customizable ally that can feel different from
run to run (lots of exciting unrand weapons to give it!), encourages the
player to get in the thick of battle themselves (it charges up the finisher
faster!), and gives not just a stat-stick, but a fancy button to press as
your level 9.

Some numbers may still need adjusting, but early playtesting feedback has
been quite positive, and I think it's a very fun spell.