Alistair's Pocket Shrapnel: new alchemy spell
A reversal of classic lead-to-gold chrysopaeia: converts a few gold
coins into silver, in an explosive reaction which can be directed toward
the enemy by using the caster's own movement. It fires in a wide beam
of 3 parallel beams (or 5 with Vehumet).

Continuing to move will cause more beams to fire in the direction of
movement, until the spell ends.

Costs actual gold, and consequently is quite powerful for its level,
but the cost limits its use.