Apply some attack flavours against dead monsters
Monster attack flavour handling was skipped if the initial hit killed a
monster's target. Most of the time this was sensible (no point in freezing
something that is already dead!), but a handful of attack flavours had
effects that reached beyond just the target they were attacking. For
example:

-Broodmothers/Obsidian Bats couldn't summon things off killing blows
-AF_VAMPIRIC monsters couldn't drain HP from killing blows (notably: the
 vampiric weapon brand *does* work on killing blows.
-Shadowghasts don't lose their invisibility on shadowstabs that kill their
 target (allowing them to reuse it immediately)
-AF_SWARM, AF_BLOODZERK, AF_BLINK similarly don't work properly.

This commit changes a handful of such attack flavours to work even if the
defender is dead.