Fix Ash bondage not updating with slot-blocking muts
Bondage level was rechecked if an equipped piece of equipment was forced
off due to a mutation, but not if you didn't have anything equipped in that
slot already. However, it should still be rechecked in those cases because
piety scales based on the number of available slots (so if, for example, a
player has every slot filled except one, and then that uncursed slot is
blocked by a mutation, they should then reach max piety).