New spell: Alistair's Walking Alembic (Level 5 Forgecraft/Alchemy)
This spell creates a potion-brewing golem with both offensive and
supportive properties.

Its melee attack vents fumes (creating short-lived poison clouds around
what it punched), and after it has performed a certain number of attacks,
it will finish brewing potions and distribute them. This affects up to 5
nearby allied monsters, and also the player (if they are within 3 tiles of
the alembic when it finishes.)

For monsters, the effect is chosen randomly among haste, might, brilliance,
or heal wounds (choosing only effects that could do something for the
monster in question - ie: no might on monsters with no attacks.).

For players, it chooses one of haste, magic, heal wounds, might, or
invisibility, at random. These effects are slightly weaker than drinking
the potion directly, but otherwise count as a 'real' potion for multiple
effects (ie: mummies cannot drink them, oni can get a cleave attack from
drinking !hw/!magic, etc.)

Now, giving the player unlimited potion effects 'for free' might seem very
strong, but I believe the combination of the randomness, as well as needing
to have this golem survive and attack a bunch of turns in a row before
getting any effect at all, can balance this out appropriately. !haste is
strong not *just* because being hasted is strong, but because you can
become hasted the moment you want to and not: a ~1-in-5 chance 5-6 turns
later.

Also, the potion connection is, I think, good flavor to make alchemy feel
more alchemy-like. A useful offensive buddy, with a unique bonus.