A variety of remixed extended tiles
All of these branches lacked their own unique tile appearances beyond a
few recolours of default tiles. With the bulk of extended running heavily
off of theme and novelty more than about replayable variety, it helps
Pan and the Hells each to have their own unique tiles for both the sake
of external vault flavour and for the flavour of each of those branches.
(I also have some ideas and work done on new Zot tiles in this capacity,
but it required each branch to technically have its own tiles first.)

 * Pandemonium stone: A mix of Denzi's wall_relief, wall_bars, and
   Porkchop's stone_dark walls recolouring Denzi's stone_grey walls.

 * Mnoleg stone variation: A mix of Denzi's wall_relief, wall_bars, and
   unrecognizably-warped old very ugly thing tile.

 * Gloorx Vloq stone: A mix of ontoclasm's Crypt metal, plus Sastreii's
   Snake stone and Depths vault floors.

 * Lom Lobon crystal: A combination of heavily edited CC0 art by NettySvit
   and cynicmusic.

 * Cerebov metal: A mix of Heavily edited CC0 art by Surt and
   unrecognizably-warped balrug tiles by Sastreii.

 * Gehenna stone: A mix of ontoclasm's Labyrinth stone and
   unrecognizably-warped pit fiend tiles by Denzi.

This also comes with some minor adjustments to a number of vaults to deploy
these tiles a bit more (for various demonic contexts), and to avoid using
them in holy_pan while that's still around. Should be a little easier to
make Pan decor any visually interesting, at least.