New start-up and clockwork bee tiles (#4102)
"The Orb of Zot on top of the leader board sprite is outdated, and the tile
size was inconsistent. Descent startup tile also needed an upgrade. The new
one shows stairs descending deeper into the dungeon."

[Commiter's notes: Changed the Descent icon to both have more depth /
height to each step and to use a red palette- gold palettes on the starting
menu are more information based compared to red changing the mode's
behaviour entirely, and it also corresponds to hatches being red.

The clockwork bee tile submission has been heavily adjusted, keeping the
gear of the current one in order to make it more obviously mechanical
versus the ally circle in-game and deploying a palette unlike the current
or the submitted tile to match other Forgecraft tiles Sastreii made.
(Sastreii will replace a large number of Forgecraft tiles as is at some
point in the future, but updating the tile to have a wind-up key seems
harmless enough to me.)

Closes #4102.]