New spell: Forge Monarch Bomb (Level 6 Fire/Forgecraft)
This spell summons a Monarch Bomb, which is a relatively weak ally with the
ability to deploy a large number of Bomblets via its melee attack (or at
range with the Launch Bomblet spell). These bomblets are mobile, but
possess no attack of their own (and cannot be ordered around).

However, when the player recasts the spell, all active bomblets and the
monarch bomb itself will detonate, dealing half-physical/half-fire damage
to all adjacent targets (and destroying themselves).

This has shades of the old Forceful Dismissal spell, but in a more specific
context. There's a certain mild 'push your luck' element with detonating
the bombs now versus waiting for more bomblets to be deployed, and setting
up and then triggering large explosions is fun! It also offers Fire magic
an optonal side path that can perform better against fire-resistant
targets, making what I think is an interesting option when approached from
either school.