Move warn_about_bad_targets, improve slightly
This adds a parameter which can be used to exclude certain targets which
cannot be harmed by the attack being warned about. (This isn't used more
widely at the moment, but there are some existing uses that probably should
be edited to account for resists.)

Also, if the same space is affected more than once, don't confusingly claim
that "Something (and 1 other bad target)" could be hit.