Fix monsters sometimes following the player into Duels (particleface)
MF_TAKING_STAIRS is set on nearby monsters when the player starts to take a
stair. This flag is used at the end of the stair delay to determine what
monsters are eligable to follow. It is cleared from non-following monsters
when the player leaves the floor, but was *not* cleared when the player's
attempt to take the stairs was interrupted (eg: by constriction, trampling,
etc.)

Oka's Duel ability enters the arena via code that handles normal stair
transitions. It does not *set* the MF_TAKING_STAIRS flag, but it still
respects it on any monsters which already had it. Thus it was possible, if
the player took the stairs, was interrupted, and then used Duel, to pull in
all the monsters that *would* have followed them across the stairs into the
Duel.

This was a rather mysterious bug to deduce and I'm still not 100% convinced
this solves all instances of Duel stowaways, but it does solve at least a
solid chunk of them.