New Xom action: light webs on fire
On one hand, it's what a lot of games that aren't Crawl would allow the
player to do anyway, and Xom already has a noticeably fiery action.

On the other hand, the idea that in the entirety of Crawl, that only Xom
can rarely decide to do this, is probably funny enough in and of itself.

Much like Snakes to Sticks, this is meant to provide a Xom action that
changes according to the branches one gets in a game- and somewhat
usefully, there's only two monsters in Spider's Nest with any fire
resistance. Unlike Snakes to Sticks, this comes with a fair bit less
complexity- it just removes all webs in sight (and frees anybody stuck in
a web), then creates fire clouds where those webs were. It provides a
little more ease of movement for the branch in a flavourful fashion without
being too intrusive, and continues the trend of Xom uniquely warping the
Dungeon's terrain on a whim.