Make quote text more visually separated from mechanical text
This adds a visual divider between mechanical descriptive text and flavor
text for most non-monster UI (monsters already have it on a seperate tab),
as well as drawing said text in a darker color.

It would be nice if there was some color between grey and darkgrey here,
and it would be *especially* nice if I could actually center the divider,
but as far as I can tell, a bunch of places that construct these
descriptions do not yet know the width of the popup they will be placed in,
so left-justified it is for now.

(While I feel many of the real-world quotes Crawl currently uses are
dubious or added out of a sense of completionism, I think the same space
could be used for ancillary flavor text or *in-world* quotes instead, and
plan to use it for such for at least a few forthcoming spells. More could
theoretically be done over time.)