New spell: Nazja's Percussive Tempering (Level 5 Forgecraft)
Another hybrid offense/defense spell, Nazja's Percussive Tempering targets
an allied creature which was created by a Forgecraft spell and applies the
Tempered buff to it as well as healing it somewhat and immediately dealing
AoE damage around that creature.

Tempering boosts the monster's HD by 4 and also increases its melee damage
by 25%.

Tempering is fairly short-lived, but importantly the spell cannot be recast
on an ally who is currently tempered. And since a large chunk of the
spell's utility is in the damage burst (and a little in the heal), the hope
is that it isn't something you auto-cast on an ally, but rather choose an
appropriate time to use.

A useful and flexible utility spell that naturally interacts with different
forgecraft allies in different ways.