New Spell: Fortress Blast (Level 6 Forgecraft)
A deliberate melee hybrid spell to further establish Forgecraft as not
*just* about making allies, but also other things conceptually related to
tangible physical structure and armou

Fortress Blast is a player-centered AoE attack spell whose damage scales
entirely based on your AC. Casting it roots the player in place for several
turns (during which they can still perform other actions), and at the end
of these turns, there is a large explosion of physical force. Spellpower
affects how quickly the blast charges up, but the power cap is deliberately
low, since this is aimed at characters with high AC who are less likely
to plausibly *reach* high power.

This is aimed, to a degree, as a sort of inverse counterpart to Manifold
Assault. Where that spell tends to be favored by EV characters (due in
part to its spell level making it harder to cast in armour), this one is
obviously aimed at AC characters. It comes with the downside of needing to
anchor you in place, but can do a lot more damage-per-MP if you're facing
many enemies. A spell to be added in a subsequent commit can also increase
the player's AC directly, providing a natural synergy for this spell (and
Animate Armour, slated to be moved to Forgecraft, naturally is boosted by
similar things, creating a cohesive sub-theme)

The damage scaling from AC is non-linear (ie: each point of AC is worth
*more* than the last) and currently caps at ~70 AC. The damage is *quite*
high with high AC, but given that it can inflict that damage at most every
few turns (along with downside), I do think it needs to be at least a
little high. (Formulas are, of course, subject to change from playtesting.)