New spell: Summon Seismosaurus Egg (Level 4 Summonings/Earth)
Intended as the other half of a replacement for Summoner losing Blazeheart
and Lightning Spire, while trying to retain some of the dynamicism that
Blazeheart added to the background.

This spell summons a dinosaur egg on a random tile between 2-3 spaces away
from the player. If the player sticks near that egg for several turns, to
imprint upon it, *while enemies are nearby*, the egg will hatch into a very
strong ally for its level. (A small heart icon will be displayed over the
egg if the conditions are right for it to hatch, making it more transparent
to players when things are sufficiently in range.)

A seismosaurus is somewhat overstatted for level 4 (to compensate for the
slowness and positioning needed to hatch one) and has the Seismic Stomp
ability, which deals smite-targeted damage to several random enemies within
4 tiles of the seismosaurus. This deal only 1/3 damage to fliers, but can
sometimes cause grounded enemies hit by it to stumble and miss their turn
outright. They will use Stomp immediately upon hatching, and can also use
it randomly thereafter.

Being a smite-targeted AoE, it plays well with all sorts of allies (ie: it
can shoot 'through' them) and positions you end up in. And I think it
naturally gives the spell a sort of progression, where when you *first* get
it, the dinosaur is very strong, but you need to put effort into holding
the line with much weaker allies in order to hatch it. And then later on,
when you have stronger allies, you can toss an egg out for some
supplementary damage behind your higher-level summons.

Initial playtesting feedback has been positive, and I think the spell is
fun to use!