New spell: Phalanx Beetle (level 6 Forgecraft)
This spell constructs a fast-moving ally whose mere presence increases the
caster's AC while it is alive, but which has a unique behavior: it will
never leave the caster's side under its own power.

If the caster moves, it will follow in lockstep with them (although it can
pivot around to reach enemies near you.) If forcibly separated (ie: via
knockback or blinking) it will prioritize returning to your side over
attacking anything else. It cannot be ordered away.

Its attack is fairly unimpressive for its level, and it cannot be thrown at
enemies, but it is durable and increases the caster's own durability (as
well as providing a natural synergy with Fortress Blast). Intended as a
generic mid-range option that can fit into almost any Forgecraft build.