New spell: Rending Blade (Level 4 Conjurations/Forgecraft)
Intended as a new Reaver capstone spell, as well as a sort of 'melee
counterpart' to Battlesphere. Continuing the reaver tradition, this is a
strong damage conjuration with significant downsides, which supports a
conjurations hybrid playstyle.

Casting this spell uses *all* of the caster's remaining MP to create a
magical blade that flies alongside you for a small number of turns.
Whenever the caster hits with a melee attack while the spell is active,
this blade will attempt to dash through a line of enemies, inflicting
irresistable damage to all of them. When the blade dies or expires, the
MP used to form it is returned to the player. (An earlier version just gave
-Cast for the duration, but I thought it was a little more interesting if
it was *possible* to restore MP during the spell's effect, and also if the
amount of remaining MP provides a small power boost.)

The blade will prioritize the most HD of enemies in can hit in one dash
without leaving the caster's sight. If no viable target exists that turn,
it will retain up to 3 charges to unleash the next time a target presents
itself. (Since the blade uses the same M_MAINTAIN_RANGE behavior that
Battlesphere does, which jitters a bit, it tends to find viable shots again
on subsequent turns.) Hitting multiple enemies in a single attack (such as
with an axe) can trigger the blade multiple times at once! This is
obviously strong, but axe conjurers are such a niche thing that I think
it's okay if it's theoretically especially good with them (after all, it's
also especially good with quick blades.)

Given that the blade *deliberately* cannot fly through the player
themselves, it tends not to function well without a little space around,
but can be quite strong with the proper terrain.

(Reavers are not an especially weak background, although not a popular one,
and anectodally, many seem to stop training conjurations after Momentum
Strike. By offering a higher-level spell that continues the hybrid
playstyle, it may make them feel a bit more attractive - at least, make
them more likely not to ignore conjurations entirely in the long-term,
in favor of melee.)