Move Battlesphere into Forgecraft, tweak, rewrite backend code
This commit moves Battlesphere from Level 5 Conjurations to Level 4
Conjurations/Forgecraft. I was a little less certain about this becoming
Forgecraft than I was the other spells I've moved, but it does allow it to
share schools with its closest relatives, Servitor and the new Rending
Blade.

At the same time, this adjusts the behavior of battlesphere slightly, and
rewrites its code almost completely. While the refactor of ca6d514eab45b7
greatly reduced the number of situations where a battlesphere looked like
it should have triggered, but did not, it didn't remove them altogether.
Battlesphere's original implementation worked by co-opting normal monster
movement/action code to try and coax the battlesphere into relocating
itself over multiple actions until it had line-of-fire to a given target.
This was quite brittle and involving an increasing amount of bespoke code
and properties to try and do so intelligently. Rather than try to add *even
more* to fix these issues, I've opted to use a different approach
altogether.

Now, when a battlesphere is looking for an alternate firing position, it
calculates what spaces it could reach from its current position within 3
movement actions (accounting for what creatures are currently in the way),
and if any of those are valid firing positions, it will just immediately
move to that location. The battlesphere was already fast enough that
'teleporting' in this way should be visually indistinguisable from doing it
the normal way, but considerably less error-prone.

At the same time, I've changed the battlesphere's targeting behavior from
'pick a target at random' to 'pick the most injured target' (as calculated
by the difference between its max and current hp). This should greatly
increase the chance of it focusing on the target the player wants it to (or
at the very least - a target it is most likely to finish off).

I've also removed the effect of spellpower on duration and number of shots
before disappearing. Doing low damage at low power is enough, I think,
rather than doing low damage *and* only lasting for a few turns. (I've set
the average number of shots per cast at around ~70 power in the old
system.)

Battlesphere's been a fairly lackluster spell in recent years, and moving
it into side schools would probably make it even less appealing without
some touch-ups, so hopefully this new version feels better to use.